% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{bookmark_time}
\alias{bookmark_time}
\alias{bookmark_time,RamMonitor,character-method}
\title{Bookmark Time (Generic)}
\usage{
bookmark_time(monitor, label)

\S4method{bookmark_time}{RamMonitor,character}(monitor, label)
}
\arguments{
\item{monitor}{A RamMonitor object}

\item{label}{A string label}
}
\description{
Bookmark Time (Generic)
}
