% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{get_current_ram_usage}
\alias{get_current_ram_usage}
\alias{get_current_ram_usage,RamMonitor-method}
\title{Get Current RAM Usage (Generic)}
\usage{
get_current_ram_usage(monitor)

\S4method{get_current_ram_usage}{RamMonitor}(monitor)
}
\arguments{
\item{monitor}{A RamMonitor object}
}
\value{
Numeric. Current RAM usage in GB.
}
\description{
Get Current RAM Usage (Generic)
}
