% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{get_max_ram_usage}
\alias{get_max_ram_usage}
\alias{get_max_ram_usage,RamMonitor-method}
\title{Get Maximum RAM Usage (Generic)}
\usage{
get_max_ram_usage(monitor)

\S4method{get_max_ram_usage}{RamMonitor}(monitor)
}
\arguments{
\item{monitor}{A RamMonitor object}
}
\value{
Numeric. Maximum RAM usage in GB.
}
\description{
Get Maximum RAM Usage (Generic)
}
