% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{get_ram_data}
\alias{get_ram_data}
\alias{get_ram_data,RamMonitor-method}
\title{Get Detailed RAM Data (Generic)}
\usage{
get_ram_data(monitor)

\S4method{get_ram_data}{RamMonitor}(monitor)
}
\arguments{
\item{monitor}{A RamMonitor object}
}
\value{
A tibble with columns \code{timestamp}, \code{memory_bytes}, \code{time}, \code{stage}
}
\description{
Retrieves the memory usage data as a tibble, including a column
indicating the "stage" of execution based on bookmarked timepoints.
}
