% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ram_usage}
\alias{plot_ram_usage}
\alias{plot_ram_usage,RamMonitor-method}
\title{Plot RAM Usage (Generic)}
\usage{
plot_ram_usage(
  monitor,
  line_size = 1.5,
  line_color = "black",
  font_size_title = 16,
  font_size_axis = 14
)

\S4method{plot_ram_usage}{RamMonitor}(
  monitor,
  line_size = 1.5,
  line_color = "black",
  font_size_title = 16,
  font_size_axis = 14
)
}
\arguments{
\item{monitor}{A RamMonitor object}

\item{line_size}{Numeric. Thickness of the RAM usage line (default: 1.5).}

\item{line_color}{Character. Color of the RAM usage line (default: "black").}

\item{font_size_title}{Numeric. Font size for the plot title (default: 16).}

\item{font_size_axis}{Numeric. Font size for the axis labels (default: 14).}
}
\value{
A ggplot2 object
}
\description{
Plot RAM Usage (Generic)
}
