% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{start_monitoring}
\alias{start_monitoring}
\title{Start Monitoring RAM}
\usage{
start_monitoring(
  pid = NULL,
  process_name = NULL,
  interval = 1,
  data_dir = tempdir()
)
}
\arguments{
\item{pid}{Process ID to monitor. Defaults to current R session.}

\item{process_name}{Optional process name. If provided, pid is ignored.}

\item{interval}{Polling interval in seconds. Default is 1.}

\item{data_dir}{Directory to store the memory log. Default is tempdir().}
}
\value{
A RamMonitor S4 object.
}
\description{
Start Monitoring RAM
}
