% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{stop_monitoring}
\alias{stop_monitoring}
\alias{stop_monitoring,RamMonitor-method}
\title{Stop Monitoring (Generic)}
\usage{
stop_monitoring(monitor, cleanup = FALSE)

\S4method{stop_monitoring}{RamMonitor}(monitor, cleanup = FALSE)
}
\arguments{
\item{monitor}{A RamMonitor object}

\item{cleanup}{Logical. If TRUE, deletes the temporary data file.
Default is FALSE.}
}
\description{
Stop Monitoring (Generic)
}
